Annotation Convertion
- 3rd party library
  - `sqlite3-amalgamation` (pre-build)
  - `csv-parser` (pre-build)
- SQLite3 in-memory database for annotation / EXIF data management
- TODO
  - add SQLite3 recommanded build options
  - update x64 cmake
