+++ .gitmodules
@@ -26,3 +26,11 @@
 	path = contrib/libpng
 	url = https://github.com/pnggroup/libpng.git
 	branch = libpng16
+[submodule "contrib/sqlite-amalgamation"]
+	path = contrib/sqlite-amalgamation
+	url = https://github.com/ClickHouse/sqlite-amalgamation
+	branch = master
+[submodule "contrib/csv-parser"]
+	path = contrib/csv-parser
+	url = https://github.com/vincentlaucsb/csv-parser
+	branch = master
