+++ CMakeLists.txt
@@ -26,6 +26,7 @@ set(CMAKE_MODULE_LINKER_FLAGS "${CMAKE_MODULE_LINKER_FLAGS} -fuse-ld=lld")
 set(SOURCES
     "${FusswegDatentools_SOURCE_DIR}/src/main.cpp"
     "${FusswegDatentools_SOURCE_DIR}/src/annot.cpp"
+    "${FusswegDatentools_SOURCE_DIR}/src/via.cpp"
     "${FusswegDatentools_SOURCE_DIR}/src/crs.cpp"
     "${FusswegDatentools_SOURCE_DIR}/src/cv.cpp"
     "${FusswegDatentools_SOURCE_DIR}/src/exif.cpp"
@@ -33,6 +34,7 @@ set(SOURCES
 )
 set(HEADERS
     "${FusswegDatentools_SOURCE_DIR}/include/annot.hpp"
+    "${FusswegDatentools_SOURCE_DIR}/include/via.hpp"
     "${FusswegDatentools_SOURCE_DIR}/include/crs.hpp"
     "${FusswegDatentools_SOURCE_DIR}/include/cv.hpp"
     "${FusswegDatentools_SOURCE_DIR}/include/exif.hpp"
@@ -61,7 +63,7 @@ configure_file(
 # -----------------------------------------------------------------------------
 enable_testing()
 add_executable(${TEST_BIN_NAME}
-    "${FusswegDatentools_SOURCE_DIR}/src/annot.cpp"
+    "${FusswegDatentools_SOURCE_DIR}/src/via.cpp"
     "${FusswegDatentools_SOURCE_DIR}/src/crs.cpp"
     "${FusswegDatentools_SOURCE_DIR}/src/exif.cpp"
     "${FusswegDatentools_SOURCE_DIR}/src/gis.cpp"
